Help from Janelyn to fix two issues causing RoboRio resets.:
1) Resets with CAN bus connected at startup.  Through an extensive process of elimination the reset issue was caused by the Reefscape motor current limits being set to 10 Amps.  After changing to 20 Amps the resets stopped.  The theory is that the motor controllers must instantaneously pull more than 10 Amps when they are starting up and somehow setting off an infinite reset loop.  This was noticed by comparing to Swerve current limits that we set to 60 Amps.
2) There was some issue with the IMU function call leading to resets.  After trying a number of things Janelyn found commands that worked and did not cause resets.

Janelyn also noticed that I had done a bad thing by using the coral arm angle .SetInverted command in the pre step callback instead of using it higher up in the file.  This was corrected.

Now all of the hand code is uncommented and should be working, except for Photon Vision that William will be working on updating.