Automatically turn on the algae wheels for intake if the buttons are pushed for algae elevator height.
Change the elevator duty cycle logic when using the gamepad 'bottom' button.  If the elevator error is greater than Elevator_Error_Bottom_Disable then set the duty cycle to the lower limit (-0.5), otherwise set to the smaller value (-0.04).  This will get the elevator down faster if an elevator height measurement reset is needed.