Disable the swerve motors when in the Disabled game state to hopefully avoid the robot steering twitch when changing to disabled.
Also set the absolute steering and translation flags to false in the disabled state (this is probably what was causing the twitch).

Added transition in Teleop Stateflow to go from Start to Algae_Score state if the Coral Limit switch is false and Algae Limit switch is true.
Also transition to Algae_Hold at startup if the Algae Limit switch is true.