**Auto path development should be complete.**
Moved the MatchTime interface calculation from GameStateChangeCallback to PreStepCallback so it would update.
The match timer is a count down instead of count up, so changed the logic from >= to <= comparisons.
Tuned the match time values to allow for as much time as possible for scoring the second L4.

Small path adjustments for center.
Keep the algae ball high (L2) until done driving to the processor so the camera can see the April tag.